+++ scratch.js
@@ -1,23 +1,4 @@
 const spacetime = require('./src/index')
 
-let s = spacetime('2017-04-03T08:00:00-0700', 'Canada/Eastern', {
-  silent: true
-})
-s.log()
-
-
-// let s = spacetime([1970, 1, 1], 'UTC')
-// console.log(new Date(s.epoch * 1000))
-//
-// s = spacetime([2054, 10, 1], 'UTC')
-// console.log(new Date(s.epoch / 1000))
-
-// new handy warning in spacetime@3.0
-// Fun fact: we're only subject-to these 'off-by-thousand' date errors between Sept 2001 and Jan 2057.
-
-// console.time('test')
-// for (let i = 0; i < 20000; i += 1) {
-//   let r = parseInt(Math.random() * 5, 10)
-//   let s = spacetime([1980 + r, r, 20 + r], 'UTC')
-// }
-// console.timeEnd('test')
+let obj = spacetime('April 12th 2008', 'Canada/Eastern').from('March 12 2018')
+console.log(obj)