+++ src/input/index.js
@@ -1,6 +1,7 @@
 'use strict';
 const strFmt = require('./strParse');
 const fns = require('../fns');
+const namedDates = require('./named-dates')
 //we have to actually parse these inputs ourselves
 //  -  can't use built-in js parser ;(
 //=========================================
@@ -40,12 +41,11 @@ const handleObject = function(s, obj) {
 };
 
 //find the epoch from different input styles
-const parseInput = (s, input, givenTz, options) => {
-  options = options || {}
+const parseInput = (s, input, givenTz) => {
   //if we've been given a epoch number, it's easy
   if (typeof input === 'number') {
     s.epoch = input;
-    if (input > 0 && input < minimumEpoch && !options.silent) {
+    if (input > 0 && input < minimumEpoch && s.silent === false) {
       console.warn('  - Warning: You are setting the date to January 1970.')
       console.warn('       -   did input seconds instead of milliseconds?')
     }
@@ -76,17 +76,28 @@ const parseInput = (s, input, givenTz, options) => {
     handleObject(s, input);
     return;
   }
+  //input as a string..
   if (typeof input !== 'string') {
     return;
   }
+  //little cleanup..
+  input = input.trim().replace(/ +/g, ' ')
+  //try some known-words, like 'now'
+  if (namedDates.hasOwnProperty(input) === true) {
+    s = namedDates[input](s)
+    return
+  }
+  //try each text-parse template, use the first good result
   for (let i = 0; i < strFmt.length; i++) {
     let m = input.match(strFmt[i].reg);
     if (m) {
-      strFmt[i].parse(s, m, givenTz, options);
+      strFmt[i].parse(s, m, givenTz);
       return;
     }
   }
-  console.warn('Warning: couldn\'t parse date-string: \'' + input + '\'')
+  if (s.silent === false) {
+    console.warn('Warning: couldn\'t parse date-string: \'' + input + '\'')
+  }
   s.epoch = null;
   s.valid = false;
   return;