+++ src/methods/diff.js
@@ -1,6 +1,7 @@
 'use strict';
 const fns = require('../fns');
-
+let doAll = () => {
+}
 //increment until dates are the same
 const climb = function(a, b, unit) {
   let i = 0;
@@ -15,7 +16,11 @@ const climb = function(a, b, unit) {
   return i;
 };
 
-const diff = (a, b, unit) => {
+const diff = function(a, b, unit) {
+  b = fns.beADate(b, a)
+  if (!unit) {
+    return doAll(a, b)
+  }
   unit = fns.normalize(unit);
   if (a.isBefore(b)) {
     return climb(a, b, unit);
@@ -23,5 +28,16 @@ const diff = (a, b, unit) => {
     //reverse it
     return climb(b, a, unit) * -1;
   }
-};
+}
+
+doAll = function(a, b) {
+  return {
+    years: diff(a, b, 'year'),
+    months: diff(a, b, 'month'),
+    weeks: diff(a, b, 'week'),
+    days: diff(a, b, 'day'),
+    hours: diff(a, b, 'hour'),
+    minutes: diff(a, b, 'minute'),
+  }
+}
 module.exports = diff;